requirements.txt
websocket-client==0.14.0
gpudb==6.2.0.10